Change text direction if Arabic is chosen
This is done in a very naive way, and doesn't work properly for
English text (which would probably not be an issue in the real
version). Currently, all the displayed text is affected. If we
want finer control, we may need to make sure each component has
access to the translation service (or at least can get the current
language).