Treat counts as numbers
This converts the icount field to an integer, rather than text.
Ideally this will be done at the back-end, but I'm not sure that
the JSON will be transferred as anything other than text. In the
meantime, this will do (the code could probably also be made more
concise). This also makes the results table field for counts
right-aligned.