Generalise host name
Using "localhost" means that Angular tries to connect to the
backend on the machine of whoever is visiting the page! Instead of
hardcoding the URL of the museum server, using this Javascript
property will let this work both on our local setups and on remote
servers. We are still assuming that the frontend and backend are
running on the same machine.