Start adapting to the new API
This shows how to only get the first ten results. Starting to put
support in place for passing other options to the backend (sorting
order and field, starting entry, count) so that the bulk of the
work (searching and sorting) takes place in the backend.
Unrelated to this, the HttpClient module is much simpler than the
older Http one, so switching to it.