Put in place structure for new display table page
We are changing the results table to only display results, rather
than do sorting and searching. These operations will be done in
the backend, to which the table component will send requests via
the data service. We are also keeping the old functionality of
displaying all results and doing the filtering/sorting from the
front-end (still using the SearchTableComponent), but making it
accessible under a new router link ("List all").

We will probably need to revisit this structure, and maybe put all
the functionality related to displaying results (the tables and the
data service) into one directory, so that we can avoid duplicating
code while keeping utility functions reasonably local (rather than
shared for the whole app). See also:
https://angular.io/guide/styleguide#overall-structural-guidelines