Clear table, don't search, if query box is empty
This was previously causing a problem when clicking the search
button without a query, as it was sending a request to an invalid
endpoint (/search/). This doesn't happen anymore. Instead, if the
button is pressed when the query (text box) is empty, the table
is cleared (same as when searching and getting no results).